+++ DragonShield/helpers/UserDefaultsKeys.swift
@@ -19,4 +19,6 @@ struct UserDefaultsKeys {
     static let backupDirectoryBookmark = "backupDirectoryBookmark"
     static let positionsVisibleColumns = "positionsVisibleColumns"
     static let positionsFontSize = "positionsFontSize"
+    /// Persist selected segment in Currencies & FX maintenance view.
+    static let currenciesFxSegment = "currenciesFxSegment"
 }